3.1 - fixed default trade logic, reconfig of rsi handling
