+++ core/action_handler.js
@@ -6,7 +6,7 @@ module.exports = function container (get, set, clear) {
   var get_tick_str = get('utils.get_tick_str')
   var map = get('map')
   return function action_handler (tick, action, rs, cb) {
-    get('logger').info('action', get_tick_str(tick.id), action.type.grey)
+    get('logger').info('action', get_tick_str(tick.id), action.type.grey, action, {feed: 'actions'})
     if (get('command') === 'run') {
       if (action.type === 'buy' || action.type === 'sell') {
         // @todo trade api
