+++ plugins/gdax/recorder.js
@@ -38,7 +38,7 @@ module.exports = function container (get, set, clear) {
             map('trade', obj)
             return obj
           })
-          //log_trades(x.name, trades)
+          log_trades(x.name, trades)
           retry()
         }
         var uri = x.rest_url + '/products/' + product.id + '/trades' + (s.recorder_id ? '?before=' + s.recorder_id : '')
