+++ utils/is_backfilled.js
@@ -2,13 +2,18 @@ var tb = require('timebucket')
 
 module.exports = function container (get, set, clear) {
   var c = get('config')
+  var rsi_backfiller = get('mappers.rsi_backfiller')
   return function is_backfilled (trades) {
     if (!trades || !trades.length) return false
     var min_time = trades.reduce(function (prev, curr) {
       return prev ? Math.min(prev, curr.time) : curr.time
     })
-    return min_time <= tb('1d')
+    var ret = min_time <= tb('1d')
       .subtract(c.backfill_days)
       .toMilliseconds()
+    if (ret) {
+      rsi_backfiller()
+    }
+    return ret
   }
 }
